+++ packages/pluggableWidgets/app-events-native/src/__tests__/AppEvents.spec.tsx
@@ -1,5 +1,4 @@
 import { actionValue } from "@mendix/piw-utils-internal";
-import { createElement } from "react";
 import { AppStateStatus } from "react-native";
 import { render } from "@testing-library/react-native";