+++ packages/pluggableWidgets/badge-native/src/__tests__/Badge.spec.tsx
@@ -1,5 +1,4 @@
 import { actionValue, dynamicValue } from "@mendix/piw-utils-internal";
-import { createElement } from "react";
 import { Platform, Text } from "react-native";
 import { fireEvent, render } from "@testing-library/react-native";