+++ packages/pluggableWidgets/bottom-sheet-native/src/BottomSheet.tsx
@@ -1,4 +1,4 @@
-import { createElement, Fragment, ReactElement } from "react";
+import { Fragment, ReactElement } from "react";
 import { BottomSheetStyle } from "./ui/Styles";
 import { CustomModalSheet } from "./components/CustomModalSheet";
 import { ExpandingDrawer } from "./components/ExpandingDrawer";