+++ packages/pluggableWidgets/gallery-native/src/Gallery.tsx
@@ -1,5 +1,5 @@
 import { all } from "deepmerge";
-import { createElement, ReactElement, useCallback, useEffect, useMemo, useState, useRef } from "react";
+import { ReactElement, useCallback, useEffect, useMemo, useState, useRef } from "react";
 import {
     executeAction,
     FilterType,