+++ packages/pluggableWidgets/signature-native/src/components/Touchable.tsx
@@ -1,4 +1,4 @@
-import { createElement, FunctionComponent } from "react";
+import { FunctionComponent } from "react";
 import {
     TouchableNativeFeedback,
     TouchableHighlight,